+++ README.md
@@ -32,10 +32,10 @@
 
 #### 3. 프로그램 종료 `APPLICATION_EXIT`
 
-- [ ] 프로그램 종료
+- [x] 프로그램 종료
 
 
-- [ ] 주문 혹은 결제가 불가능한 경우 그 이유를 보여 주고, 다시 주문 혹은 결제가 가능하도록 해야 한다.
+- [x] 주문 혹은 결제가 불가능한 경우 그 이유를 보여 주고, 다시 주문 혹은 결제가 가능하도록 해야 한다.
 
 # java-chicken-2019
