+++ src/main/java/util/Util.java
@@ -1,32 +1,14 @@
 package util;
 
-import java.util.Arrays;
-import java.util.List;
-
 public class Util {
 
     public static String removeSpace(String input) {
         return input.replaceAll(Regex.SPACE.regex, Regex.NO_SPACE.regex);
     }
 
-    public static String removeDelimiters(String input) {
-        return input.replace(Regex.SQUARE_BRACKETS_START.regex, Regex.NO_SPACE.regex)
-                .replace(Regex.SQUARE_BRACKETS_END.regex, Regex.NO_SPACE.regex);
-    }
-
-    public static List<String> splitByComma(String input) {
-        return Arrays.asList(Util.removeSpace(input).split(Regex.COMMA.regex));
-    }
-
-    public static List<String> formatProductInfo(String input) {
-        return Util.splitByComma(Util.removeDelimiters(Util.removeSpace(input)));
-    }
-
 
     private enum Regex {
-        SPACE(" "), NO_SPACE(""),
-        SQUARE_BRACKETS_START("["), SQUARE_BRACKETS_END("]"),
-        COMMA(",");
+        SPACE(" "), NO_SPACE("");
 
         private final String regex;
